fix: move return field names into the package
