+++ india_compliance/gst_returns/fields/gstr3b.py
@@ -1,0 +1,1 @@
+"""GSTR-3B constants. None today (raw literals in code); will be filled in later PRs."""
