+++ india_compliance/gst_returns/fields/ims.py
@@ -1,0 +1,3 @@
+"""IMS constants. Shares gstr2.Category; own fields are raw literals in mappers, extracted in future PRs."""
+
+from .gstr2 import Category  # noqa: F401
